"feat: Module existing structures as functions"
