+++ src/components/add-new-object/generate.new.object.js
@@ -2,7 +2,7 @@ import { generateModalFragment } from "../modals/modal.manager.js";
 import { addNewClassModal } from "../modals/templates-modal/add.class.modal.js";
 import { addNewTeacherModal } from "../modals/templates-modal/add.teacher.modal.js";
 import { addNewStudentModal } from "../modals/templates-modal/add.student.modal.js";
-import { saveDataToLocalStorage } from "../../functions.js";
+import { manageSaveEvents ,showAddingModal } from "../../functions.js";
 
 export function generateNewObjectModal() {
   const pEl = document.getElementById("add-text");
@@ -23,20 +23,9 @@ export function generateNewObjectModal() {
     default:
       break;
   }
-  const dynamicModal = new bootstrap.Modal(
-    document.getElementById("dynamicModal")
-  );
-  dynamicModal.show();
-  const saveChangesButtons = [
-    ...document.getElementsByClassName("saveChanges"),
-  ];
+  
+  showAddingModal()
 
-  saveChangesButtons.forEach((button) => {
-    button.addEventListener("click", (event) => {
-      const dataType = event.target.id;
-      console.log("CHECKPOINT", dataType);
-      saveDataToLocalStorage(dataType);
-    });
-  });
+  manageSaveEvents()
   
 }
