+++ src/pages/classes/components/class.card.generator.js
@@ -14,7 +14,7 @@ export function generateClassCardsHTML() {
                 <a href="#" class="text-danger"><i class="fas fa-trash-alt"></i></a>
               </div>
               <h5 class="card-title text-danger">${classEl.className}</h5>
-              <h6 class="card-subtitle mb-2 text-warning">${classEl.teacher}</h6>
+              <h6 class="card-subtitle mb-2 text-warning">${classEl.teachers[0].teacherName}</h6>
               <p class="card-text">${classEl.data}</p>
               <div class= " d-flex justify-content-around">
                 <a href="#" class="card-link text-info">Students</a>
